helper-cli: Rename the `common` package to `utils`
This better fits the conventions of the majority of code base.
