+++ helper-cli/src/main/kotlin/commands/ListCopyrightsCommand.kt
@@ -28,10 +28,10 @@ import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.options.required
 import com.github.ajalt.clikt.parameters.types.file
 
-import org.ossreviewtoolkit.helper.common.PackageConfigurationOption
-import org.ossreviewtoolkit.helper.common.createProvider
-import org.ossreviewtoolkit.helper.common.processAllCopyrightStatements
-import org.ossreviewtoolkit.helper.common.readOrtResult
+import org.ossreviewtoolkit.helper.utils.PackageConfigurationOption
+import org.ossreviewtoolkit.helper.utils.createProvider
+import org.ossreviewtoolkit.helper.utils.processAllCopyrightStatements
+import org.ossreviewtoolkit.helper.utils.readOrtResult
 import org.ossreviewtoolkit.model.Identifier
 import org.ossreviewtoolkit.model.config.CopyrightGarbage
 import org.ossreviewtoolkit.model.config.orEmpty
