+++ helper-cli/src/main/kotlin/commands/packageconfig/SortCommand.kt
@@ -24,8 +24,8 @@ import com.github.ajalt.clikt.parameters.arguments.argument
 import com.github.ajalt.clikt.parameters.arguments.convert
 import com.github.ajalt.clikt.parameters.types.file
 
-import org.ossreviewtoolkit.helper.common.sortEntries
-import org.ossreviewtoolkit.helper.common.write
+import org.ossreviewtoolkit.helper.utils.sortEntries
+import org.ossreviewtoolkit.helper.utils.write
 import org.ossreviewtoolkit.model.config.PackageConfiguration
 import org.ossreviewtoolkit.model.readValue
 import org.ossreviewtoolkit.utils.common.expandTilde
