+++ helper-cli/src/main/kotlin/commands/packagecuration/SplitCommand.kt
@@ -25,10 +25,10 @@ import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.options.required
 import com.github.ajalt.clikt.parameters.types.file
 
-import org.ossreviewtoolkit.helper.common.formatComment
-import org.ossreviewtoolkit.helper.common.getSplitCurationFile
-import org.ossreviewtoolkit.helper.common.readPackageCurations
-import org.ossreviewtoolkit.helper.common.writeAsYaml
+import org.ossreviewtoolkit.helper.utils.formatComment
+import org.ossreviewtoolkit.helper.utils.getSplitCurationFile
+import org.ossreviewtoolkit.helper.utils.readPackageCurations
+import org.ossreviewtoolkit.helper.utils.writeAsYaml
 import org.ossreviewtoolkit.utils.common.expandTilde
 
 internal class SplitCommand : CliktCommand(
