+++ helper-cli/src/main/kotlin/commands/repoconfig/ImportPathExcludesCommand.kt
@@ -26,16 +26,16 @@ import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.options.required
 import com.github.ajalt.clikt.parameters.types.file
 
-import org.ossreviewtoolkit.helper.common.VcsUrlMapping
-import org.ossreviewtoolkit.helper.common.getRepositoryPath
-import org.ossreviewtoolkit.helper.common.getRepositoryPaths
-import org.ossreviewtoolkit.helper.common.importPathExcludes
-import org.ossreviewtoolkit.helper.common.mergePathExcludes
-import org.ossreviewtoolkit.helper.common.orEmpty
-import org.ossreviewtoolkit.helper.common.readOrtResult
-import org.ossreviewtoolkit.helper.common.replacePathExcludes
-import org.ossreviewtoolkit.helper.common.sortPathExcludes
-import org.ossreviewtoolkit.helper.common.write
+import org.ossreviewtoolkit.helper.utils.VcsUrlMapping
+import org.ossreviewtoolkit.helper.utils.getRepositoryPath
+import org.ossreviewtoolkit.helper.utils.getRepositoryPaths
+import org.ossreviewtoolkit.helper.utils.importPathExcludes
+import org.ossreviewtoolkit.helper.utils.mergePathExcludes
+import org.ossreviewtoolkit.helper.utils.orEmpty
+import org.ossreviewtoolkit.helper.utils.readOrtResult
+import org.ossreviewtoolkit.helper.utils.replacePathExcludes
+import org.ossreviewtoolkit.helper.utils.sortPathExcludes
+import org.ossreviewtoolkit.helper.utils.write
 import org.ossreviewtoolkit.model.OrtResult
 import org.ossreviewtoolkit.model.config.RepositoryConfiguration
 import org.ossreviewtoolkit.model.readValue
